Fix locale provider usage
