+++ app/layout.tsx
@@ -1,41 +1,17 @@
 import type { Metadata } from "next";
-import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
-import { QueryClientWrapper } from "@/components/providers/query-client-provider";
 import { LocaleProvider } from "@/components/providers/locale-provider";
-
-const geistSans = Geist({
-  variable: "--font-geist-sans",
-  subsets: ["latin"],
-});
-
-const geistMono = Geist_Mono({
-  variable: "--font-geist-mono",
-  subsets: ["latin"],
-});
+import { LayoutWithLocale } from "@/components/providers/layout-with-locale";
 
 export const metadata: Metadata = {
   title: "Create Next App",
   description: "Generated by create next app",
 };
 
-function InnerLayout({ children, locale = "pt-BR" }: { children: React.ReactNode; locale?: string }) {
-  "use client";
-  return (
-    <html lang={locale}>
-      <body className={`${geistSans.variable} ${geistMono.variable} antialiased`}>
-        <QueryClientWrapper>{children}</QueryClientWrapper>
-      </body>
-    </html>
-  );
-}
-
 export default function RootLayout({ children }: { children: React.ReactNode }) {
   return (
     <LocaleProvider>
-      <LocaleProvider.Consumer>
-        {({ locale }) => <InnerLayout locale={locale}>{children}</InnerLayout>}
-      </LocaleProvider.Consumer>
+      <LayoutWithLocale>{children}</LayoutWithLocale>
     </LocaleProvider>
   );
 }
