+++ locales/en-US.json
@@ -53,7 +53,7 @@
   "charts.negative": "Negative",
   "charts.totalBalance": "Total Balance",
   "charts.savingsRate": "Savings Rate",
-  "charts.annualProjection": "Annual Projection"
+  "charts.annualProjection": "Annual Projection",
   "sharedModal.title": "Shared Expenses/Income",
   "sharedModal.totalIncome": "Total Income",
   "sharedModal.totalExpense": "Total Expenses",
