+++ locales/pt-BR.json
@@ -53,7 +53,7 @@
   "charts.negative": "Negativo",
   "charts.totalBalance": "Saldo Total",
   "charts.savingsRate": "Taxa de Economia",
-  "charts.annualProjection": "Projeção Anual"
+  "charts.annualProjection": "Projeção Anual",
   "sharedModal.title": "Despesas/Receitas Divididas",
   "sharedModal.totalIncome": "Total Receitas",
   "sharedModal.totalExpense": "Total Despesas",
